Cleaned up output
As it says in kama-ante.js, the font doesn't render properly while you're typing it if it doesn't have a space at the beginning. But that space was showing up in the final output (pini e linja), even though that didn't need the space. So here, I removed the space when you submit it so that you can copy and paste more easily.